requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
psycopg2-binary==2.9.3
